Add files via upload
we worked on the analysis of amr diab's songs, and we are almost done with the analysis.
we  determined the  longest and the shortest songs. we looked for the most frequent appearing words in his songs and plotted graphs about this. we also plotted graphs to see the evolution of the songs length across decades.